feature/REG-74 pass hub key on startup
