+++ .travis.yml
@@ -1,5 +1,7 @@
 language: node_js
 node_js: 10.18.1
+services:
+  - mongodb
 install:
   - npm install
   - npm run setup
@@ -10,8 +12,8 @@ script:
   - npm run browserTest:run
 deploy:
   provider: npm
-  email: '$NPM_EMAIL'
-  api_token: '$NPM_API_TOKEN'
+  email: "$NPM_EMAIL"
+  api_token: "$NPM_API_TOKEN"
   ffile_glob: true
   file: build/*
   skip_cleanup: true
