+++ package.json
@@ -4,7 +4,7 @@
   "description": "The javascript library client that can be used to manage decentralised identities on energy web chains",
   "main": "index.js",
   "scripts": {
-    "test": "./node_modules/.bin/lerna run --ignore=@ew-did-registry/did-hub-client --ignore=@ew-did-registry/did-hub test ",
+    "test": "./node_modules/.bin/lerna run --ignore=@ew-did-registry/did-hub test ",
     "test-rpc": "./node_modules/.bin/run-with-testrpc -m \"candy maple cake sugar pudding cream honey rich smooth crumble sweet treat\" --port 8544 --accounts 20 --networkId=9 --gasLimit=10000000 \"npm test\"",
     "setup": "./node_modules/.bin/lerna bootstrap --hoist --no-ci",
     "build": "./node_modules/.bin/lerna run compile && ./node_modules/.bin/gulp",
