+++ packages/did-hub-client/package.json
@@ -21,6 +21,7 @@
     "@decentralized-identity/hub-common-js": "0.1.1",
     "@decentralized-identity/hub-node-core": "0.1.4",
     "@decentralized-identity/hub-sdk-js": "0.1.2",
+    "@ew-did-registry/did-hub": "^0.0.1",
     "@ew-did-registry/keys": "^1.0.0",
     "ec-key": "0.0.4",
     "eciesjs": "^0.3.5",
