+++ packages/did-hub-client/src/crypto/index.ts
@@ -2,3 +2,4 @@ export * from './DidCryptoSuit';
 export * from './DidPrivateKey';
 export * from './DidPublicKey';
 export * from './PublicKey';
+export * from './CommitSigner';
