+++ packages/did-hub/start.js
@@ -1,0 +1,3 @@
+const { runHub } = require('./hub');
+
+runHub();
